#2 #4 created inital project
Since there were some problems with permissions and
the initial commit, this is the placeholder pull
request for the initial commit.

Spring app with 4 modules created:
- app
- domain
- webApi
- githubAdapter

Minimum dependencies added to each and a simple test
to check that everything interacts correctly.

Additionally added editor config with file width
and tab spaces.