#13 integrated sonar cloud
The sonar service can be run through gradle.
Will upload a report to sonar cloud and wait
for the response. Fails if the report is not
passing.

Added this to the Github workflow to fail the
pipeline if sonar is not happy.

To use it either add a ~/.gradle/gradle.properties
file in your home directory with the properties:
- systemProp.sonar.host.url=https://sonarcloud.io
- systemProp.sonar.token=<token>
and add the personal token or define a env var

called SONAR_TOKEN