#31 #39 repositories are now fetchable
through a usecase in the domain repositories are now
fetchable. This usecase also allows for a caching
layer to be put inbetween.

The addition of this made me change the workflowRuns
usecase so that it uses the repositories to fetch all
runs for each repositories instead of a hardcoded repo.

This implementation of the repository fetching also
includes recursion in regards to paging so that all
repositories of the specified organization will be
fetched.