#52 improved testing environment
Added a docker compose that will spin up a
prometheus instance and a grafana instance
with the correct everything set up correctly
so that a spring app running on 8080 will
feed them data.

The grafana dashboards are commited to the
repo and will be loaded if the docker compose
file is run but updates to the dashboards will
not be saved correctly unless done so manually.

Also added the first integration test to properly
test from the githubapi all the way to the
prometheus endpoint.