#56 Cleaned up Schedulers for both work runs and jobs (#61)
split up the nested loops in the scheduler classes to seperate use case classes to make the code easier to read

not 100% happy with the getAllJobsUseCase calling the getAllWorkflowRunsUseCase. not sure if its wrong(?) or just feels weird. but its the only way I could think to not have duplicate code.