#73 inverted exporting order
The prometheus exporter now pulls from the domain to export to the actuator
endpoint. Every metric has its own exporter which takes care of calling
the domain and also talking to micrometer.

Before the actual Scheduling was beeing started for every exporter in the
Scheduler, this is not simplified by having all exporters implement the
ScheduledExporter interface and then injecting them in the Scheduler.