#81 Add cache eviction back in to the system (#88)
added the @CacheEvict annotation to the run function which gets called from the scheduler in the prometheus exporter
we realize this does make the caching bleed into the exporter from the GH adapter but there doesn't seem to be a way around that currently. a potential future solution to this would be have a webhook that cache evicts on say a new workflow run starting. and then this change is the easiest to refactor rather then a more in depth mod