properly using the RestClient (#97)
#75 Refactor Exception handling as well as Authorization code and make the
RestClient handle that

In this PR I changed all of this functionality to work through the rest client
directly instead of doing these things through method created by us.

## Authorization

Before we would have a method to add the correct headers to every request we
decided to make

Now these headers are added automatically through a Interceptor removing that
responsiblity from the user of the RestClient. This also means that every
request using the "defaultRestClient" now will be authorized with the GithubAppToken.

To be able to fetch the First token there is a separate RestClient called
"tokenFetcherRestClient". This client in turn has a interceptor that will add
the correct JWT token to the request.

## Exception handling

Before there were two different methods that would check for body parsing error
and translate them to our internal exceptions.

Now this is also replaced by a status handler that catches 401's and 403's and
throws a GithubRequestWasUnauthenticatedException. Compared to the previous
methods this status handler does not cover every single excpetion that was beeing
caught before. This has been done on purpuse since some of them should have never
been caught in the first place.