#91 fixed pmd and checkstyle config files
Previously the pmd config file was in the checkstyle directory which
made no sense so now its moved out and just sits in the config
directory with the checkstyle file aswell.

Since this actually activated pmd for the first time I also handled
all the warnings given by pmd.

One oddity is that both the checkstyle as well as pmd config files have
different paths depending on which module tries to look for them. This means
that submodules find the config files in ../config and the main module in just
./config. I dont know whether there is a better way of doing this.