#110 added pagination function (#115)
Created a generic pagination function that allows for very easy
pagination of any different request. The idea is that any reqest
to an api endpoint will return some mappingclass which can be
mapped to a list of domain objects. The generic function I created
takes in this function, the initial response entity and then
follows the pagination as far as it can and collects all the
responses into one big list of domain objects.

Consider that this pagination functionality will have to be
reworked once #112 is beeing implemented as this functionality
will follow the pagination blindly until the end.