#40 self hosted runners metric
Added feature that exposes the number of self hosted runners for a
organization. Since runners can also be configured only on a specific
repository, separate queries are made to both the org as well as
every repository to get all the runners their status and os.

The status is a combination of the busy and status fields since
not every combination of those is actually possible.

Reason for the flattening of the repository/org runners is becuase more
then the state and number of runners is not interesting to us.