#166 fixed selfhosted runner os parsing error (#171)
The os string returned by the Github api turned out to be more unpredictable then expected.
This meant that a more complex setup was needed to convert the os string to the OperatingSystem
enum.

Converting the string to enum works by trying to find any of a list of value in the returned os
string. This list can be extended from the application.yaml through the APP_GITHUB_<OS>_SELFHOSTED_RUNNER_OS_KEYWORDS
environemnt variable.

Additionally, instead of throwing an Exception when the parsing didn't succeed I added a
UNKNOWN variant to the enum and log the os string for later reference.