#170 unstucking the rate limit status
Before the problem was that the status could only update when requests where beeing made meaning
that in the case where all requests were stopped because the status changed to CRITICAL for
example it would never update.

Now I have created a control task that runs every so often (default 5 mins) and checks if
requests have been made in that timeframe. If that is not the case then it lowers the status
by one. (Consider that since the status is an optional and empty in the case of a limit await
this lowering by one might actually not do anything)